chore: allow to test webhook on creation
